Eliminando todos los archivos que se han dejado de utilizar
