+++ frontend/server/nginx.rewrites
@@ -88,7 +88,6 @@ rewrite ^/problem/random/karel/?$ /problems/random/karel.php last;
 rewrite ^/problem/([a-zA-Z0-9_+-]+)/edit/?$ /problems/edit.php?problem=$1 last;
 rewrite ^/problem/([a-zA-Z0-9_+-]+)/stats/?$ /problems/stats.php?problem_alias=$1 last;
 rewrite ^/profile/?$ /profile/index.php last;
-rewrite ^/profile/edit/?$ /profile/edit.php last;
 rewrite ^/profile/([a-zA-Z0-9_+.-]+)/?$ /profile/index.php?username=$1 last;
 rewrite ^/rank/?$ /rank/users.php last;
 rewrite ^/rank/authors/?$ /rank/authors.php last;
