--- frontend/templates/user.basicedit.tpl
@@ -1,5 +1,0 @@
-{extends file="user.edit.tpl"}
-{block name="basic-content"}
-	<div id="user-basic-edit"></div>
-	{js_include entrypoint="user_basic_edit"}
-{/block}
