+++ frontend/www/js/omegaup/components/user/ManageIdentities.vue
@@ -1,22 +1,16 @@
 <template>
-  <div class="panel panel-default">
-    <div class="panel-heading">
-      <h2 class="panel-title">{{ T.profileManageIdentities }}</h2>
-    </div>
-    <div class="panel-body add-identity-panel">
-      <form
-        class="form add-identity-form"
-        @submit.prevent="$emit('add-identity', { username, password })"
-      >
-        <div class="form-group">
-          <label>{{ T.wordsIdentity }}</label>
-          <span
-            aria-hidden="true"
-            class="glyphicon glyphicon-info-sign"
-            data-placement="top"
-            data-toggle="tooltip"
+  <div>
+    <form
+      class="form add-identity-form"
+      @submit.prevent="$emit('add-identity', { username, password })"
+    >
+      <div class="form-group">
+        <label class="w-100">
+          {{ T.wordsIdentity }}
+          <font-awesome-icon
             :title="T.profileAddIdentitiesTooltip"
-          ></span>
+            icon="info-circle"
+          ></font-awesome-icon>
           <input
             v-model="username"
             autocomplete="off"
@@ -24,9 +18,11 @@
             size="20"
             type="text"
           />
-        </div>
-        <div class="form-group">
-          <label>{{ T.loginPassword }}</label>
+        </label>
+      </div>
+      <div class="form-group">
+        <label class="w-100">
+          {{ T.loginPassword }}
           <input
             v-model="password"
             autocomplete="off"
@@ -34,31 +30,31 @@
             size="20"
             type="password"
           />
-        </div>
-        <div class="form-group pull-right">
-          <button class="btn btn-primary" type="submit">
-            {{ T.wordsAddIdentity }}
-          </button>
-        </div>
-      </form>
-      <div v-if="identities.length == 0">
-        <div class="empty-category">
-          {{ T.profileIdentitiesEmpty }}
-        </div>
+        </label>
       </div>
-      <table v-else class="table table-striped table-over">
-        <thead>
-          <tr>
-            <th>{{ T.wordsIdentity }}</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr v-for="identity in identities" :key="identity.username">
-            <td>{{ identity.username }}</td>
-          </tr>
-        </tbody>
-      </table>
+      <div class="form-group text-right">
+        <button class="btn btn-primary" type="submit">
+          {{ T.wordsAddIdentity }}
+        </button>
+      </div>
+    </form>
+    <div v-if="identities.length == 0">
+      <div class="empty-category">
+        {{ T.profileIdentitiesEmpty }}
+      </div>
     </div>
+    <table v-else class="table table-striped table-over">
+      <thead>
+        <tr>
+          <th>{{ T.wordsIdentity }}</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr v-for="identity in identities" :key="identity.username">
+          <td>{{ identity.username }}</td>
+        </tr>
+      </tbody>
+    </table>
   </div>
 </template>
 
@@ -67,17 +63,27 @@ import { Vue, Component, Prop } from 'vue-property-decorator';
 import { types } from '../../api_types';
 import T from '../../lang';
 
-@Component
-export default class UserManageIdentities extends Vue {
+import {
+  FontAwesomeIcon,
+  FontAwesomeLayers,
+  FontAwesomeLayersText,
+} from '@fortawesome/vue-fontawesome';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { library } from '@fortawesome/fontawesome-svg-core';
+library.add(fas);
+
+@Component({
+  components: {
+    'font-awesome-icon': FontAwesomeIcon,
+    'font-awesome-layers': FontAwesomeLayers,
+    'font-awesome-layers-text': FontAwesomeLayersText,
+  },
+})
+export default class ManageIdentities extends Vue {
   @Prop() identities!: types.Identity[];
+
   T = T;
-  username: string = '';
-  password: string = '';
+  username = '';
+  password = '';
 }
 </script>
-
-<style>
-th.align-right {
-  text-align: right;
-}
-</style>
